extend remote storage service to external interface
