+++ app/src/main/java/org/stepik/android/adaptive/api/profile/model/ProfileRequest.kt
@@ -1,0 +1,5 @@
+package org.stepik.android.adaptive.api.profile.model
+
+import org.stepik.android.adaptive.data.model.Profile
+
+class ProfileRequest(private val profile: Profile)
