+++ app/src/main/java/org/stepik/android/adaptive/api/profile/model/ProfileResponse.kt
@@ -1,0 +1,9 @@
+package org.stepik.android.adaptive.api
+
+import org.stepik.android.adaptive.data.model.Profile
+import org.stepik.android.adaptive.data.model.User
+
+class ProfileResponse(@JvmField val profiles: List<Profile>?) {
+    val profile: Profile?
+        get() = profiles?.firstOrNull()
+}
